Fix bug where miner could finish early if main proc was busy with IO.
The default select case could be hit if the channel was not being
actively read from by the main proc because it was busy saving another
result to the disk.

The default case is not needed anymore since the introduction of the
context to control cancellation of the miner. It can safely be removed,
which fixes the bug.