Added new ReadOnionAddress function to public package API.
The new function ReadOnionAddress reads a directory containing the 3
files generated from the SaveOnionAddress function to construct to the
files back into an OnionAddress struct so it can continue to be worked
with programatically.

It actually only needs 2 of the 3 files that are created: the public key
file and the private key file. The host name can be calculated using
that data so is not used when reading.

A variant of this new function called ReadOnionAddressFS has also been
added, which does the same thing but accepts an fs.FS interface so you
can work with abstracted file systems if required.